+++ SimpleAnimation/ProblemAnimations.m
@@ -249,7 +249,7 @@ copiedView.backgroundColor = self.flipTooMuchView.backgroundColor;
 
 - (IBAction)jerkyRotatePressed:(id)sender 
 {
-    static int animationCounter = 0;
+    static NSInteger animationCounter = 0;
 
     [UIView animateWithDuration:1.0
                      animations:^
@@ -268,7 +268,7 @@ copiedView.backgroundColor = self.flipTooMuchView.backgroundColor;
 
 - (IBAction)correctRotatePressed:(id)sender 
 {
-    static int animationCounter = 0;
+    static NSInteger animationCounter = 0;
     
     [UIView animateWithDuration:1.0
                           delay:0
